refactor(mllp-client): naming and structure for public release
Naming consistency — prefix all public identifiers with MllpClient so
they read unambiguously when imported in isolation:

  ClientTlsOptions   -> MllpClientTlsOptions
  ClientErrorCode    -> MllpClientErrorCode

Acknowledgment keeps its package-agnostic name because it describes
the entity itself, not a MllpClient-bound configuration.

Internal/public separation — split client.ts (700+ lines) into a
small public layer plus a focused internal/ directory:

  src/
    index.ts            public exports
    client.ts           MllpClient class + public option types  (~245 lines)
    errors.ts           MllpClientError + MllpClientErrorCode
    acknowledgment.ts   Acknowledgment + parseAck/throwOnNak
    internal/
      socket.ts         openConnection + mapSocketError + destroySocket
      exchange.ts       exchange + writeFrame + readFirstFrame + encodeOrThrow
      deadline.ts       Deadline + createDeadline
      run-once.ts       makeRunOnce

The public package surface is unchanged in shape — index.ts re-exports
the same set of symbols — but the file layout makes it easier to
evolve internals without touching the public file and makes it obvious
which modules are not part of the API.
