+++ packages/mllp-client/src/errors.ts
@@ -2,15 +2,23 @@ import { MllpError } from "@glion/mllp-transport";
 import type { MllpErrorOptions } from "@glion/mllp-transport";
 
 /**
- * Machine-readable codes for client-side transport failures.
+ * Machine-readable codes for `MllpClient` transport failures.
  *
  * These cover everything that can go wrong between the moment
- * `MllpClient.send()` is called and the moment a complete ACK frame has
- * been decoded — i.e., everything that is not an HL7v2 application-level
- * NAK. NAK responses (MSA-1 ∈ {AE, AR, CE, CR}) surface as `AckException`
- * subclasses thrown directly from `@glion/ack`, not as transport errors.
+ * `MllpClient.send()` is called and the moment a complete ACK frame
+ * has been decoded — i.e. everything that is not an HL7v2
+ * application-level NAK. NAK responses (MSA-1 ∈ {AE, AR, CE, CR})
+ * surface as `AckException` subclasses thrown directly from
+ * `@glion/ack`, not as transport errors.
  */
-export const ClientErrorCode = {
+export const MllpClientErrorCode = {
+  /**
+   * The caller passed a payload that could not be MLLP-encoded —
+   * typically because it was not a `string` or `Uint8Array`. Thrown
+   * before any socket is opened, so no connection is made and no
+   * resources need cleanup.
+   */
+  INVALID_INPUT: "INVALID_INPUT",
   /** TCP/TLS connection could not be established (refused, DNS, routing). */
   CONNECTION_REFUSED: "CONNECTION_REFUSED",
   /** The peer closed the connection before a complete ACK arrived. */
@@ -21,38 +29,35 @@ export const ClientErrorCode = {
   MALFORMED_FRAME: "MALFORMED_FRAME",
   /** A frame was received but could not be parsed as a valid HL7v2 ACK. */
   MALFORMED_ACK: "MALFORMED_ACK",
-  /**
-   * The caller passed a payload that could not be MLLP-encoded — typically
-   * because it was not a `string` or `Uint8Array`. Thrown before any
-   * socket is opened, so no connection is made and no resources need
-   * cleanup.
-   */
-  INVALID_INPUT: "INVALID_INPUT",
 } as const;
 
-export type ClientErrorCode =
-  (typeof ClientErrorCode)[keyof typeof ClientErrorCode];
+export type MllpClientErrorCode =
+  (typeof MllpClientErrorCode)[keyof typeof MllpClientErrorCode];
 
 /**
  * Error thrown by `MllpClient.send()` for transport-level failures.
  *
  * Extends {@link MllpError} so callers can catch any MLLP error —
- * server- or client-side — with a single `instanceof MllpError` check.
- * The {@link ClientErrorCode | code} field discriminates the failure
- * mode for typed handling.
+ * server- or client-side — with a single `instanceof MllpError`
+ * check. The {@link MllpClientErrorCode | code} field discriminates
+ * the failure mode for typed handling.
  *
  * Application-level NAK responses are **not** thrown as
- * `MllpClientError`; they are thrown as `AckException` subclasses from
- * `@glion/ack`. A typical `catch` block handles both:
+ * `MllpClientError`; they are thrown as `AckException` subclasses
+ * from `@glion/ack`. A typical `catch` block handles both:
  *
  * ```ts
  * import { AckException } from "@glion/ack";
  * import { MllpClientError } from "@glion/mllp-client";
  *
- * try { await client.send(message); }
- * catch (error) {
- *   if (error instanceof AckException)        { /* application NAK *\/ }
- *   else if (error instanceof MllpClientError) { /* transport error *\/ }
+ * try {
+ *   await client.send(message);
+ * } catch (error) {
+ *   if (error instanceof AckException) {
+ *     // Application-level NAK.
+ *   } else if (error instanceof MllpClientError) {
+ *     // Transport-level failure.
+ *   }
  * }
  * ```
  */
@@ -60,14 +65,14 @@ export class MllpClientError extends MllpError {
   /**
    * Narrowed view of the inherited `MllpError.code` field. `declare`
    * keeps the runtime value (set by the base constructor) while
-   * exposing the discriminated `ClientErrorCode` union to TypeScript
-   * — callers can `switch` on `error.code` with full exhaustiveness
-   * checking.
+   * exposing the discriminated {@link MllpClientErrorCode} union to
+   * TypeScript — callers can `switch` on `error.code` with full
+   * exhaustiveness checking.
    */
-  declare readonly code: ClientErrorCode;
+  declare readonly code: MllpClientErrorCode;
 
   constructor(
-    code: ClientErrorCode,
+    code: MllpClientErrorCode,
     message: string,
     options?: MllpErrorOptions
   ) {
