+++ packages/mllp-client/src/index.ts
@@ -4,18 +4,18 @@
  * The package intentionally exposes a small set of symbols:
  *
  * - {@link MllpClient} — the class application code instantiates.
- * - {@link MllpClientOptions}, {@link ClientTlsOptions} — configuration shapes
- *   passed to the constructor.
+ * - {@link MllpClientOptions}, {@link MllpClientTlsOptions} — configuration
+ *   shapes passed to the constructor.
  * - {@link Acknowledgment} — the structured ACK type returned by `client.send()`
  *   on success.
- * - {@link MllpClientError}, {@link ClientErrorCode} — the typed transport error
- *   and its discriminant codes.
+ * - {@link MllpClientError}, {@link MllpClientErrorCode} — the typed transport
+ *   error and its discriminant codes.
  *
  * Application-level NAK exceptions (`AckApplicationError`,
  * `AckApplicationReject`, `AckCommitError`, `AckCommitReject`, etc.)
  * live in `@glion/ack` and are imported directly from there. The
- * client throws those classes, but does not re-export them — keeping
- * a single source of truth for the exception hierarchy and making it
+ * client throws those classes but does not re-export them — keeping a
+ * single source of truth for the exception hierarchy and making it
  * obvious that `@glion/ack` is the authoritative module.
  *
  * @module
@@ -27,7 +27,7 @@
 // Client
 // -------------
 export { MllpClient } from "./client";
-export type { ClientTlsOptions, MllpClientOptions } from "./client";
+export type { MllpClientOptions, MllpClientTlsOptions } from "./client";
 
 // -------------
 // Acknowledgment
@@ -37,4 +37,4 @@ export type { Acknowledgment } from "./acknowledgment";
 // -------------
 // Errors
 // -------------
-export { ClientErrorCode, MllpClientError } from "./errors";
+export { MllpClientError, MllpClientErrorCode } from "./errors";
